Fix Multiple Release Failure
Added a fix for when `lock.Release` is called multiple times. Track the status of the lock state with `lock.released` and implement locking in `lock.Release` to ensure we don't have a race condition when releasing a lock. If the lock has already been released then don't attempt to do so a second time and return an error indicating that the lock was already released. Note that functions for `Lock` had to be changed to use pointers so that we can safely use the mutex and `lock.released` field.